Fixed warnings treated as errors.

+++ model/godleyTable.h
@@ -155,7 +155,7 @@ namespace minsky
       return data[row][col];
     }
     bool cellInTable(int row, int col) const
-    {return row>=0 && row<rows() && col>=0 && col<cols();}
+    {return row>=0 && size_t(row)<rows() && col>=0 && size_t(col)<cols();}
     string getCell(unsigned row, unsigned col) const {
       if (row<rows() && col<cols())
         return cell(row,col);
